requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
Appium-Python-Client==1.0.2
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
importlib-metadata==3.3.0
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
